demo/django/requirements.txt
Django~=5.0.1
django-ninja~=1.1.0
uvicorn~=0.25.0
# fixme: should be -e git..
-e ../..
